+++ swarm/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.43.0 - unreleased
+
+- Remove `ConnectionHandlerUpgrErr::Timer` variant.
+  This variant was never constructed and thus dead code.
+  See [PR 3605].
+
+[PR 3605]: https://github.com/libp2p/rust-libp2p/pull/3605
+
 # 0.42.0
 
 - Allow `NetworkBehaviour`s to manage connections.